Move: 학습 관련 엔티티 위치 변경
