+++ biengual-core/src/main/java/com/biengual/core/domain/entity/learning/UserLearningHistoryEntity.java
@@ -1,4 +1,4 @@
-package com.biengual.core.domain.entity.userlearninghistory;
+package com.biengual.core.domain.entity.learning;
 
 import com.biengual.core.domain.entity.BaseEntity;
 import jakarta.persistence.*;
